Network.AWS.OpsWorksCM.CreateServer

Description

Creates and immedately starts a new server. The server is ready to use when it is in the HEALTHY state. By default, you can create a maximum of 10 servers.

This operation is asynchronous.

A LimitExceededException is thrown when you have created the maximum number of servers (10). A ResourceAlreadyExistsException is thrown when a server with the same name already exists in the account. A ResourceNotFoundException is thrown when you specify a backup ID that is not valid or is for a backup that does not exist. A ValidationException is thrown when parameters of the request are not valid.

If you do not specify a security group by adding the SecurityGroupIds parameter, AWS OpsWorks creates a new security group. The default security group opens the Chef server to the world on TCP port 443. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22.

By default, the Chef Server is accessible from any IP address. We recommend that you update your security group rules to allow access from known IP addresses and address ranges only. To edit security group rules, open Security Groups in the navigation pane of the EC2 management console.

Creating a Request

createServer

Arguments

:: Text
csServerName

-> Text
csInstanceProfileARN

-> Text
csInstanceType

-> Text
csServiceRoleARN

-> CreateServer

Creates a value of CreateServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csEngineVersion - The major release version of the engine that you want to use. Values depend on the engine that you choose.

csDisableAutomatedBackup - Enable or disable scheduled backups. Valid values are true or false . The default value is true .

csSecurityGroupIds - A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds . If you do not specify this parameter, AWS OpsWorks for Chef Automate creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

csAssociatePublicIPAddress - Associate a public IP address with a server that you are launching. Valid values are true or false . The default value is true .

csSubnetIds - The IDs of subnets in which to launch the server EC2 instance. Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled. EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled. For more information about supported Amazon EC2 platforms, see Supported Platforms .

csKeyPair - The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

csBackupId - If you specify this field, AWS OpsWorks for Chef Automate creates the server by using the backup represented by BackupId.

csEngine - The configuration management engine to use. Valid values include Chef .

csEngineModel - The engine model, or option. Valid values include Single .

csEngineAttributes - Optional engine attributes on a specified server. Attributes accepted in a createServer request: * CHEF_PIVOTAL_KEY : A base64-encoded RSA private key that is not stored by AWS OpsWorks for Chef. This private key is required to access the Chef API. When no CHEF_PIVOTAL_KEY is set, one is generated and returned in the response. * CHEF_DELIVERY_ADMIN_PASSWORD : The password for the administrative user in the Chef Automate GUI. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_DELIVERY_ADMIN_PASSWORD is set, one is generated and returned in the response.

csPreferredMaintenanceWindow - The start time for a one-hour period each week during which AWS OpsWorks for Chef Automate performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM . The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information. Example: Mon:08:00 , which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

csPreferredBackupWindow - The start time for a one-hour period during which AWS OpsWorks for Chef Automate backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats: * HH:MM for daily backups * DDD:HH:MM for weekly backups The specified time is in coordinated universal time (UTC). The default value is a random, daily start time. Example: 08:00 , which represents a daily start time of 08:00 UTC. Example: Mon:08:00 , which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

csBackupRetentionCount - The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks for Chef Automate deletes the oldest backups if this number is exceeded. The default value is 1 .

csServerName - The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

csInstanceProfileARN - The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

csInstanceType - The Amazon EC2 instance type to use. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large . Valid values are t2.medium , m4.large , or m4.2xlarge .

csServiceRoleARN - The service role that the AWS OpsWorks for Chef Automate service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-stuff/latest/service-role-creation.yaml. This template creates a CloudFormation stack that includes the service role that you need.

data CreateServer

See: createServer smart constructor.

Request Lenses

csEngineVersion :: Lens' CreateServer (Maybe Text)

The major release version of the engine that you want to use. Values depend on the engine that you choose.

csDisableAutomatedBackup :: Lens' CreateServer (Maybe Bool)

Enable or disable scheduled backups. Valid values are true or false . The default value is true .

csSecurityGroupIds :: Lens' CreateServer [Text]

A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds . If you do not specify this parameter, AWS OpsWorks for Chef Automate creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

csAssociatePublicIPAddress :: Lens' CreateServer (Maybe Bool)

Associate a public IP address with a server that you are launching. Valid values are true or false . The default value is true .

csSubnetIds :: Lens' CreateServer [Text]

The IDs of subnets in which to launch the server EC2 instance. Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled. EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled. For more information about supported Amazon EC2 platforms, see Supported Platforms .

csKeyPair :: Lens' CreateServer (Maybe Text)

The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

csBackupId :: Lens' CreateServer (Maybe Text)

If you specify this field, AWS OpsWorks for Chef Automate creates the server by using the backup represented by BackupId.

csEngine :: Lens' CreateServer (Maybe Text)

The configuration management engine to use. Valid values include Chef .

csEngineModel :: Lens' CreateServer (Maybe Text)

The engine model, or option. Valid values include Single .

csEngineAttributes :: Lens' CreateServer [EngineAttribute]

Optional engine attributes on a specified server. Attributes accepted in a createServer request: * CHEF_PIVOTAL_KEY : A base64-encoded RSA private key that is not stored by AWS OpsWorks for Chef. This private key is required to access the Chef API. When no CHEF_PIVOTAL_KEY is set, one is generated and returned in the response. * CHEF_DELIVERY_ADMIN_PASSWORD : The password for the administrative user in the Chef Automate GUI. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_DELIVERY_ADMIN_PASSWORD is set, one is generated and returned in the response.

csPreferredMaintenanceWindow :: Lens' CreateServer (Maybe Text)

The start time for a one-hour period each week during which AWS OpsWorks for Chef Automate performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM . The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information. Example: Mon:08:00 , which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

csPreferredBackupWindow :: Lens' CreateServer (Maybe Text)

The start time for a one-hour period during which AWS OpsWorks for Chef Automate backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats: * HH:MM for daily backups * DDD:HH:MM for weekly backups The specified time is in coordinated universal time (UTC). The default value is a random, daily start time. Example: 08:00 , which represents a daily start time of 08:00 UTC. Example: Mon:08:00 , which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

csBackupRetentionCount :: Lens' CreateServer (Maybe Natural)

The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks for Chef Automate deletes the oldest backups if this number is exceeded. The default value is 1 .

csServerName :: Lens' CreateServer Text

The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

csInstanceProfileARN :: Lens' CreateServer Text

The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

csInstanceType :: Lens' CreateServer Text

The Amazon EC2 instance type to use. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large . Valid values are t2.medium , m4.large , or m4.2xlarge .

csServiceRoleARN :: Lens' CreateServer Text

The service role that the AWS OpsWorks for Chef Automate service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-stuff/latest/service-role-creation.yaml. This template creates a CloudFormation stack that includes the service role that you need.

Destructuring the Response

createServerResponse

Arguments

:: Int
csrsResponseStatus

-> CreateServerResponse

Creates a value of CreateServerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csrsServer - The server that is created by the request.

csrsResponseStatus - -- | The response status code.

data CreateServerResponse

See: createServerResponse smart constructor.

Response Lenses

csrsServer :: Lens' CreateServerResponse (Maybe Server)

The server that is created by the request.

csrsResponseStatus :: Lens' CreateServerResponse Int

- | The response status code.