Network.AWS.OpsWorksCM.DeleteBackup

Description

Deletes a backup. You can delete both manual and automated backups. This operation is asynchronous.

An InvalidStateException is thrown when a backup deletion is already in progress. A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is thrown when parameters of the request are not valid.

Creating a Request

deleteBackup

Arguments

:: Text
dbBackupId

-> DeleteBackup

Creates a value of DeleteBackup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbBackupId - The ID of the backup to delete. Run the DescribeBackups command to get a list of backup IDs. Backup IDs are in the format ServerName-yyyyMMddHHmmssSSS .

data DeleteBackup

See: deleteBackup smart constructor.

Request Lenses

dbBackupId :: Lens' DeleteBackup Text

The ID of the backup to delete. Run the DescribeBackups command to get a list of backup IDs. Backup IDs are in the format ServerName-yyyyMMddHHmmssSSS .

Destructuring the Response

deleteBackupResponse

Arguments

:: Int
dbrsResponseStatus

-> DeleteBackupResponse

Creates a value of DeleteBackupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbrsResponseStatus - -- | The response status code.

data DeleteBackupResponse

See: deleteBackupResponse smart constructor.

Response Lenses

dbrsResponseStatus :: Lens' DeleteBackupResponse Int

- | The response status code.