Network.AWS.OpsWorksCM.DeleteServer

Description

Deletes the server and the underlying AWS CloudFormation stack (including the server's EC2 instance). When you run this command, the server state is updated to DELETING . After the server is deleted, it is no longer returned by DescribeServer requests. If the AWS CloudFormation stack cannot be deleted, the server cannot be deleted.

This operation is asynchronous.

An InvalidStateException is thrown when a server deletion is already in progress. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

deleteServer

Arguments

:: Text
dsServerName

-> DeleteServer

Creates a value of DeleteServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsServerName - The ID of the server to delete.

data DeleteServer

See: deleteServer smart constructor.

Request Lenses

dsServerName :: Lens' DeleteServer Text

The ID of the server to delete.

Destructuring the Response

deleteServerResponse

Arguments

:: Int
dsrsResponseStatus

-> DeleteServerResponse

Creates a value of DeleteServerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsResponseStatus - -- | The response status code.

data DeleteServerResponse

See: deleteServerResponse smart constructor.

Response Lenses

dsrsResponseStatus :: Lens' DeleteServerResponse Int

- | The response status code.