Network.AWS.OpsWorksCM.DescribeNodeAssociationStatus

Description

Returns the current status of an existing association or disassociation request.

A ResourceNotFoundException is thrown when no recent association or disassociation request with the specified token is found, or when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

describeNodeAssociationStatus

Arguments

:: Text
dnasNodeAssociationStatusToken

-> Text
dnasServerName

-> DescribeNodeAssociationStatus

Creates a value of DescribeNodeAssociationStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnasNodeAssociationStatusToken - Undocumented member.

dnasServerName - The name of the server from which to disassociate the node.

data DescribeNodeAssociationStatus

See: describeNodeAssociationStatus smart constructor.

Request Lenses

dnasNodeAssociationStatusToken :: Lens' DescribeNodeAssociationStatus Text

Undocumented member.

dnasServerName :: Lens' DescribeNodeAssociationStatus Text

The name of the server from which to disassociate the node.

Destructuring the Response

describeNodeAssociationStatusResponse

Arguments

:: Int
dnasrsResponseStatus

-> NodeAssociationStatus
dnasrsNodeAssociationStatus

-> DescribeNodeAssociationStatusResponse

Creates a value of DescribeNodeAssociationStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnasrsResponseStatus - -- | The response status code.

dnasrsNodeAssociationStatus - The status of the association or disassociation request. Possible values: * SUCCESS : The association or disassociation succeeded. * FAILED : The association or disassociation failed. * IN_PROGRESS : The association or disassociation is still in progress.

data DescribeNodeAssociationStatusResponse

See: describeNodeAssociationStatusResponse smart constructor.

Response Lenses

dnasrsResponseStatus :: Lens' DescribeNodeAssociationStatusResponse Int

- | The response status code.

dnasrsNodeAssociationStatus :: Lens' DescribeNodeAssociationStatusResponse NodeAssociationStatus

The status of the association or disassociation request. Possible values: * SUCCESS : The association or disassociation succeeded. * FAILED : The association or disassociation failed. * IN_PROGRESS : The association or disassociation is still in progress.