Network.AWS.OpsWorksCM.DescribeServers

Description

Lists all configuration management servers that are identified with your account. Only the stored results from Amazon DynamoDB are returned. AWS OpsWorks for Chef Automate does not query other services.

This operation is synchronous.

A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

describeServers :: DescribeServers

Creates a value of DescribeServers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssServerName - Describes the server with the specified ServerName.

dssNextToken - NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeServers again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null . Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

dssMaxResults - To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

data DescribeServers

See: describeServers smart constructor.

Request Lenses

dssServerName :: Lens' DescribeServers (Maybe Text)

Describes the server with the specified ServerName.

dssNextToken :: Lens' DescribeServers (Maybe Text)

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeServers again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null . Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

dssMaxResults :: Lens' DescribeServers (Maybe Natural)

To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Destructuring the Response

describeServersResponse

Arguments

:: Int
dssrsResponseStatus

-> DescribeServersResponse

Creates a value of DescribeServersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssrsServers - Contains the response to a DescribeServers request.

dssrsNextToken - NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeServers again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null . Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

dssrsResponseStatus - -- | The response status code.

data DescribeServersResponse

See: describeServersResponse smart constructor.

Response Lenses

dssrsServers :: Lens' DescribeServersResponse [Server]

Contains the response to a DescribeServers request.

dssrsNextToken :: Lens' DescribeServersResponse (Maybe Text)

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeServers again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null . Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

dssrsResponseStatus :: Lens' DescribeServersResponse Int

- | The response status code.