Network.AWS.OpsWorksCM.DisassociateNode

Description

Disassociates a node from a Chef server, and removes the node from the Chef server's managed nodes. After a node is disassociated, the node key pair is no longer valid for accessing the Chef API. For more information about how to associate a node, see AssociateNode .

A node can can only be disassociated from a server that is in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

disassociateNode

Arguments

:: Text
dnServerName

-> Text
dnNodeName

-> DisassociateNode

Creates a value of DisassociateNode with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnEngineAttributes - Engine attributes used for disassociating the node. Attributes accepted in a DisassociateNode request: * CHEF_ORGANIZATION : The Chef organization with which the node was associated. By default only one organization named default can exist.

dnServerName - The name of the server from which to disassociate the node.

dnNodeName - The name of the Chef client node.

data DisassociateNode

See: disassociateNode smart constructor.

Request Lenses

dnEngineAttributes :: Lens' DisassociateNode [EngineAttribute]

Engine attributes used for disassociating the node. Attributes accepted in a DisassociateNode request: * CHEF_ORGANIZATION : The Chef organization with which the node was associated. By default only one organization named default can exist.

dnServerName :: Lens' DisassociateNode Text

The name of the server from which to disassociate the node.

dnNodeName :: Lens' DisassociateNode Text

The name of the Chef client node.

Destructuring the Response

disassociateNodeResponse

Arguments

:: Int
dnrsResponseStatus

-> DisassociateNodeResponse

Creates a value of DisassociateNodeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnrsNodeAssociationStatusToken - Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the disassociation request.

dnrsResponseStatus - -- | The response status code.

data DisassociateNodeResponse

See: disassociateNodeResponse smart constructor.

Response Lenses

dnrsNodeAssociationStatusToken :: Lens' DisassociateNodeResponse (Maybe Text)

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the disassociation request.

dnrsResponseStatus :: Lens' DisassociateNodeResponse Int

- | The response status code.