Network.AWS.OpsWorksCM.RestoreServer

Description

Restores a backup to a server that is in a CONNECTION_LOST , HEALTHY , RUNNING , UNHEALTHY , or TERMINATED state. When you run RestoreServer, the server's EC2 instance is deleted, and a new EC2 instance is configured. RestoreServer maintains the existing server endpoint, so configuration management of the server's client devices (nodes) should continue to work.

This operation is asynchronous.

An InvalidStateException is thrown when the server is not in a valid state. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

restoreServer

Arguments

:: Text
rsBackupId

-> Text
rsServerName

-> RestoreServer

Creates a value of RestoreServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsKeyPair - The name of the key pair to set on the new EC2 instance. This can be helpful if the administrator no longer has the SSH key.

rsInstanceType - The type of the instance to create. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large . Valid values are t2.medium , m4.large , and m4.2xlarge . If you do not specify this parameter, RestoreServer uses the instance type from the specified backup.

rsBackupId - The ID of the backup that you want to use to restore a server.

rsServerName - The name of the server that you want to restore.

data RestoreServer

See: restoreServer smart constructor.

Request Lenses

rsKeyPair :: Lens' RestoreServer (Maybe Text)

The name of the key pair to set on the new EC2 instance. This can be helpful if the administrator no longer has the SSH key.

rsInstanceType :: Lens' RestoreServer (Maybe Text)

The type of the instance to create. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large . Valid values are t2.medium , m4.large , and m4.2xlarge . If you do not specify this parameter, RestoreServer uses the instance type from the specified backup.

rsBackupId :: Lens' RestoreServer Text

The ID of the backup that you want to use to restore a server.

rsServerName :: Lens' RestoreServer Text

The name of the server that you want to restore.

Destructuring the Response

restoreServerResponse

Arguments

:: Int
rsrsResponseStatus

-> RestoreServerResponse

Creates a value of RestoreServerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsrsResponseStatus - -- | The response status code.

data RestoreServerResponse

See: restoreServerResponse smart constructor.

Response Lenses

rsrsResponseStatus :: Lens' RestoreServerResponse Int

- | The response status code.