Network.AWS.OpsWorksCM.StartMaintenance

Description

Manually starts server maintenance. This command can be useful if an earlier maintenance attempt failed, and the underlying cause of maintenance failure has been resolved. The server is in an UNDER_MAINTENANCE state while maintenance is in progress.

Maintenance can only be started on servers in HEALTHY and UNHEALTHY states. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

startMaintenance

Arguments

:: Text
smServerName

-> StartMaintenance

Creates a value of StartMaintenance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smServerName - The name of the server on which to run maintenance.

data StartMaintenance

See: startMaintenance smart constructor.

Request Lenses

smServerName :: Lens' StartMaintenance Text

The name of the server on which to run maintenance.

Destructuring the Response

startMaintenanceResponse

Arguments

:: Int
smrsResponseStatus

-> StartMaintenanceResponse

Creates a value of StartMaintenanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smrsServer - Contains the response to a StartMaintenance request.

smrsResponseStatus - -- | The response status code.

data StartMaintenanceResponse

See: startMaintenanceResponse smart constructor.

Response Lenses

smrsServer :: Lens' StartMaintenanceResponse (Maybe Server)

Contains the response to a StartMaintenance request.

smrsResponseStatus :: Lens' StartMaintenanceResponse Int

- | The response status code.