Network.AWS.OpsWorksCM.Types

Description

Service Configuration

opsWorksCM :: Service

API version 2016-11-01 of the Amazon OpsWorks for Chef Automate SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the provided request parameters are not valid.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource cannot be created because it already exists.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

This occurs when the provided nextToken is not valid.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource does not exist, or access was denied.

_InvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is in a state that does not allow you to perform a specified action.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The limit of servers or backups has been reached.

BackupStatus

data BackupStatus

Constructors

BSDeleting

BSFailed

BSInProgress

BSOK

BackupType

data BackupType

Constructors

Automated

Manual

MaintenanceStatus

data MaintenanceStatus

Constructors

MSFailed

MSSuccess

NodeAssociationStatus

data NodeAssociationStatus

The status of the association or disassociation request.

Possible values:

SUCCESS : The association or disassociation succeeded.

FAILED : The association or disassociation failed.

IN_PROGRESS : The association or disassociation is still in progress.

Constructors

NASFailed

NASInProgress

NASSuccess

ServerStatus

data ServerStatus

Constructors

BackingUp

ConnectionLost

Creating

Deleting

Failed

Healthy

Modifying

Restoring

Running

Setup

Terminated

UnderMaintenance

Unhealthy

AccountAttribute

data AccountAttribute

Stores account attributes.

See: accountAttribute smart constructor.

accountAttribute :: AccountAttribute

Creates a value of AccountAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aaUsed - The current usage, such as the current number of servers that are associated with the account.

aaMaximum - The maximum allowed value.

aaName - The attribute name. The following are supported attribute names. * ServerLimit: The number of current serversmaximum number of servers allowed. By default, you can have a maximum of 10 servers. * ManualBackupLimit: The number of current manual backupsmaximum number of backups allowed. By default, you can have a maximum of 50 manual backups saved.

aaUsed :: Lens' AccountAttribute (Maybe Int)

The current usage, such as the current number of servers that are associated with the account.

aaMaximum :: Lens' AccountAttribute (Maybe Int)

The maximum allowed value.

aaName :: Lens' AccountAttribute (Maybe Text)

The attribute name. The following are supported attribute names. * ServerLimit: The number of current serversmaximum number of servers allowed. By default, you can have a maximum of 10 servers. * ManualBackupLimit: The number of current manual backupsmaximum number of backups allowed. By default, you can have a maximum of 50 manual backups saved.

Backup

data Backup

Describes a single backup.

See: backup smart constructor.

backup :: Backup

Creates a value of Backup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bEngineVersion - The engine version that is obtained from the server when the backup is created.

bServiceRoleARN - The service role ARN that is obtained from the server when the backup is created.

bStatus - The status of a backup while in progress.

bInstanceProfileARN - The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

bSecurityGroupIds - The security group IDs that are obtained from the server when the backup is created.

bStatusDescription - An informational message about backup status.

bServerName - The name of the server from which the backup was made.

bSubnetIds - The subnet IDs that are obtained from the server when the backup is created.

bKeyPair - The key pair that is obtained from the server when the backup is created.

bCreatedAt - The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

bBackupId - The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

bEngine - The engine type that is obtained from the server when the backup is created.

bInstanceType - The instance type that is obtained from the server when the backup is created.

bEngineModel - The engine model that is obtained from the server when the backup is created.

bPreferredMaintenanceWindow - The preferred maintenance period that is obtained from the server when the backup is created.

bUserARN - The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

bPreferredBackupWindow - The preferred backup period that is obtained from the server when the backup is created.

bS3LogURL - The Amazon S3 URL of the backup's log file.

bS3DataSize - This field is deprecated and is no longer used.

bBackupARN - The ARN of the backup.

bS3DataURL - This field is deprecated and is no longer used.

bDescription - A user-provided description for a manual backup. This field is empty for automated backups.

bBackupType - The backup type. Valid values are automated or manual .

bToolsVersion - The version of AWS OpsWorks for Chef Automate-specific tools that is obtained from the server when the backup is created.

bEngineVersion :: Lens' Backup (Maybe Text)

The engine version that is obtained from the server when the backup is created.

bServiceRoleARN :: Lens' Backup (Maybe Text)

The service role ARN that is obtained from the server when the backup is created.

bStatus :: Lens' Backup (Maybe BackupStatus)

The status of a backup while in progress.

bInstanceProfileARN :: Lens' Backup (Maybe Text)

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

bSecurityGroupIds :: Lens' Backup [Text]

The security group IDs that are obtained from the server when the backup is created.

bStatusDescription :: Lens' Backup (Maybe Text)

An informational message about backup status.

bServerName :: Lens' Backup (Maybe Text)

The name of the server from which the backup was made.

bSubnetIds :: Lens' Backup [Text]

The subnet IDs that are obtained from the server when the backup is created.

bKeyPair :: Lens' Backup (Maybe Text)

The key pair that is obtained from the server when the backup is created.

bCreatedAt :: Lens' Backup (Maybe UTCTime)

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

bBackupId :: Lens' Backup (Maybe Text)

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

bEngine :: Lens' Backup (Maybe Text)

The engine type that is obtained from the server when the backup is created.

bInstanceType :: Lens' Backup (Maybe Text)

The instance type that is obtained from the server when the backup is created.

bEngineModel :: Lens' Backup (Maybe Text)

The engine model that is obtained from the server when the backup is created.

bPreferredMaintenanceWindow :: Lens' Backup (Maybe Text)

The preferred maintenance period that is obtained from the server when the backup is created.

bUserARN :: Lens' Backup (Maybe Text)

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

bPreferredBackupWindow :: Lens' Backup (Maybe Text)

The preferred backup period that is obtained from the server when the backup is created.

bS3LogURL :: Lens' Backup (Maybe Text)

The Amazon S3 URL of the backup's log file.

bS3DataSize :: Lens' Backup (Maybe Int)

This field is deprecated and is no longer used.

bBackupARN :: Lens' Backup (Maybe Text)

The ARN of the backup.

bS3DataURL :: Lens' Backup (Maybe Text)

This field is deprecated and is no longer used.

bDescription :: Lens' Backup (Maybe Text)

A user-provided description for a manual backup. This field is empty for automated backups.

bBackupType :: Lens' Backup (Maybe BackupType)

The backup type. Valid values are automated or manual .

bToolsVersion :: Lens' Backup (Maybe Text)

The version of AWS OpsWorks for Chef Automate-specific tools that is obtained from the server when the backup is created.

EngineAttribute

data EngineAttribute

A name and value pair that is specific to the engine of the server.

See: engineAttribute smart constructor.

engineAttribute :: EngineAttribute

Creates a value of EngineAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eaValue - The value of the engine attribute.

eaName - The name of the engine attribute.

eaValue :: Lens' EngineAttribute (Maybe Text)

The value of the engine attribute.

eaName :: Lens' EngineAttribute (Maybe Text)

The name of the engine attribute.

Server

data Server

Describes a configuration management server.

See: server smart constructor.

server :: Server

Creates a value of Server with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sEngineVersion - The engine version of the server. Because Chef is the engine available in this release, the valid value for EngineVersion is 12 .

sServiceRoleARN - The service role ARN used to create the server.

sDisableAutomatedBackup - Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

sStatus - The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

sInstanceProfileARN - The instance profile ARN of the server.

sSecurityGroupIds - The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

sAssociatePublicIPAddress - Associate a public IP address with a server that you are launching.

sServerName - The name of the server.

sSubnetIds - The subnet IDs specified in a CreateServer request.

sKeyPair - The key pair associated with the server.

sCreatedAt - Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

sServerARN - The ARN of the server.

sEngine - The engine type of the server. The valid value in this release is Chef .

sMaintenanceStatus - The status of the most recent server maintenance run. Shows SUCCESS or FAILED .

sInstanceType - The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

sEngineModel - The engine model of the server. The valid value in this release is Single .

sEngineAttributes - The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks for Chef Automate; they are returned only as part of the result of createServer(). Attributes returned in a createServer response: * CHEF_PIVOTAL_KEY : A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API. * CHEF_STARTER_KIT : A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

sPreferredMaintenanceWindow - The preferred maintenance period specified for the server.

sPreferredBackupWindow - The preferred backup period specified for the server.

sStatusReason - Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

sEndpoint - A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io

sCloudFormationStackARN - The ARN of the CloudFormation stack that was used to create the server.

sBackupRetentionCount - The number of automated backups to keep.

sEngineVersion :: Lens' Server (Maybe Text)

The engine version of the server. Because Chef is the engine available in this release, the valid value for EngineVersion is 12 .

sServiceRoleARN :: Lens' Server (Maybe Text)

The service role ARN used to create the server.

sDisableAutomatedBackup :: Lens' Server (Maybe Bool)

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

sStatus :: Lens' Server (Maybe ServerStatus)

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

sInstanceProfileARN :: Lens' Server (Maybe Text)

The instance profile ARN of the server.

sSecurityGroupIds :: Lens' Server [Text]

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

sAssociatePublicIPAddress :: Lens' Server (Maybe Bool)

Associate a public IP address with a server that you are launching.

sServerName :: Lens' Server (Maybe Text)

The name of the server.

sSubnetIds :: Lens' Server [Text]

The subnet IDs specified in a CreateServer request.

sKeyPair :: Lens' Server (Maybe Text)

The key pair associated with the server.

sCreatedAt :: Lens' Server (Maybe UTCTime)

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

sServerARN :: Lens' Server (Maybe Text)

The ARN of the server.

sEngine :: Lens' Server (Maybe Text)

The engine type of the server. The valid value in this release is Chef .

sMaintenanceStatus :: Lens' Server (Maybe MaintenanceStatus)

The status of the most recent server maintenance run. Shows SUCCESS or FAILED .

sInstanceType :: Lens' Server (Maybe Text)

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

sEngineModel :: Lens' Server (Maybe Text)

The engine model of the server. The valid value in this release is Single .

sEngineAttributes :: Lens' Server [EngineAttribute]

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks for Chef Automate; they are returned only as part of the result of createServer(). Attributes returned in a createServer response: * CHEF_PIVOTAL_KEY : A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API. * CHEF_STARTER_KIT : A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

sPreferredMaintenanceWindow :: Lens' Server (Maybe Text)

The preferred maintenance period specified for the server.

sPreferredBackupWindow :: Lens' Server (Maybe Text)

The preferred backup period specified for the server.

sStatusReason :: Lens' Server (Maybe Text)

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

sEndpoint :: Lens' Server (Maybe Text)

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io

sCloudFormationStackARN :: Lens' Server (Maybe Text)

The ARN of the CloudFormation stack that was used to create the server.

sBackupRetentionCount :: Lens' Server (Maybe Int)

The number of automated backups to keep.

ServerEvent

data ServerEvent

An event that is related to the server, such as the start of maintenance or backup.

See: serverEvent smart constructor.

serverEvent :: ServerEvent

Creates a value of ServerEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

seLogURL - The Amazon S3 URL of the event's log file.

seServerName - The name of the server on or for which the event occurred.

seCreatedAt - The time when the event occurred.

seMessage - A human-readable informational or status message.

seLogURL :: Lens' ServerEvent (Maybe Text)

The Amazon S3 URL of the event's log file.

seServerName :: Lens' ServerEvent (Maybe Text)

The name of the server on or for which the event occurred.

seCreatedAt :: Lens' ServerEvent (Maybe UTCTime)

The time when the event occurred.

seMessage :: Lens' ServerEvent (Maybe Text)

A human-readable informational or status message.