Network.AWS.OpsWorksCM.UpdateServer

Description

Updates settings for a server.

This operation is synchronous.

Creating a Request

updateServer

Arguments

:: Text
usServerName

-> UpdateServer

Creates a value of UpdateServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usDisableAutomatedBackup - Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

usPreferredMaintenanceWindow - Undocumented member.

usPreferredBackupWindow - Undocumented member.

usBackupRetentionCount - Sets the number of automated backups that you want to keep.

usServerName - The name of the server to update.

data UpdateServer

See: updateServer smart constructor.

Request Lenses

usDisableAutomatedBackup :: Lens' UpdateServer (Maybe Bool)

Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

usPreferredMaintenanceWindow :: Lens' UpdateServer (Maybe Text)

Undocumented member.

usPreferredBackupWindow :: Lens' UpdateServer (Maybe Text)

Undocumented member.

usBackupRetentionCount :: Lens' UpdateServer (Maybe Int)

Sets the number of automated backups that you want to keep.

usServerName :: Lens' UpdateServer Text

The name of the server to update.

Destructuring the Response

updateServerResponse

Arguments

:: Int
usrsResponseStatus

-> UpdateServerResponse

Creates a value of UpdateServerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usrsServer - Contains the response to a UpdateServer request.

usrsResponseStatus - -- | The response status code.

data UpdateServerResponse

See: updateServerResponse smart constructor.

Response Lenses

usrsServer :: Lens' UpdateServerResponse (Maybe Server)

Contains the response to a UpdateServer request.

usrsResponseStatus :: Lens' UpdateServerResponse Int

- | The response status code.